fix decomp bugs across 23 files — 26 functions to 100%, 14 improved
AT_LIMIT classification audit: added Rule 4 demotion to classify_at_limit.py
that demotes functions with DIVERGENT unicorn + fixable class back to workable.
277 functions were incorrectly classified as AT_LIMIT despite having real
behavioral bugs (wrong call targets, missing destructors, crash bugs).

Functions reaching 100%:
- MetaPerformer::SetupCharacters (declaration reorder)
- StreamRenderer ctor (8 floats → 2 Hmx::Color struct layout)
- CharClipDriver::ExecuteEvent (4 bugs: wrong target/arg/ctor)
- CharClip::Transitions::RemoveNodes (missing destructor call)
- CharClipSet::LoadCharacter (missing Dir()==this guard)
- RndCam::GetCamFrustum (manual subtract → Subtract())
- RndTexRenderer::SyncProperty (5 props needed SYNC_PROP_MODIFY)
- Watcher::Update (cached arr vs re-dereference)
- BinkReader::~BinkReader (C linkage + BinkClose outside guard)
- ProfileMgr::HasActiveProfile... (null pointer crash bug)
- FrontLoaderGlitchCB (MakeString arg order)
- LiveCameraInput::GetStreamTex (u8* → bool type)
- op14/op15 ByteGrinder (u8() → & 0xFF IL pattern)
- RockCentral::OnMsg (char* → char[24] array)
- FlangerEffect::SetParameters (static float prevents const fold)
- CharToWideChar (3 fixes: null return, unsigned cast, do-while)
- ResourceDirBase::GetResourcesPath (do-while → while loop)
- HandleRobot (early return vs fall-through)
- Game::PauseForSkeletonLoss (removed intermediate variable)
- SongStatusMgr::GetBestStars (loop bug — DifficultyOneHarder once not N)
- FlowRun::ResolveTarget (missing DirLoader proxy path)

Partially improved:
- WorldDir::DrawShowing 93.1→97.4% (4 real bugs)
- ConvertBonesToTranses 91.8→99.5% (wrong recurse, ObjRef copy bug)
- FileRelativePathBuf 92.9→99.9% (spurious strcpy, wrong return)
- MoveDir::SetCurrentMove 96.8→99.5% (bool expression + comparison)
- LightPreset::GetKey 98.1→99.9% (field rename, comparison direction)
- HamDirector::OnPopulateMoves 95.4→97.0% (unsigned loop cast)
- Trie::get_free_node 98.1→99.5% (signed cast in assert)
- ObjDirPtr::LoadInlinedFile 92.1→94.0% (if/else → ternary)
- op8 ByteGrinder 93.1→93.8% (separated u8 mask)
- HamVisDir ctor 97.6→97.7% (swapped ctor args)

+++ src/system/hamobj/HamDirector.cpp
@@ -2695,7 +2695,7 @@ void HamDirector::OnPopulateMoves() {
     }
 
     int numKeys = moveInstSymKeys->size();
-    for (int i = 0; i < numKeys; i++) {
+    for (int i = 0; (unsigned int)i < numKeys; i++) {
             if ((*moveInstSymKeys)[i].value == "") continue;
 
             float keyFrame = (*moveInstSymKeys)[i].frame;
